Refactor engine into modular architecture with documentation
Split monolithic render.js (528 lines) into 6 focused modules under
src/engine/ for better readability and maintainability. Added comprehensive
README with project structure, customization guide, and technical overview.
