+++ src/components/App.js
@@ -1,15 +1,10 @@
-import React, { Component } from 'react';
-import { TryOn } from './TryOn'
+import React from 'react';
+import { TryOn } from './TryOn';
 
-export default class App extends Component {
-
-  render() {
+export default function App() {
     return (
-      <div className="App">
-        <TryOn onReady={this.ready}/>
-      </div>
-
+        <div className="App">
+            <TryOn />
+        </div>
     );
-  }
-  
 }
